RecordDataを_typesに引き離し
